requirements.txt
requests==2.31.0
pandas==2.2.1
xlsxwriter==3.1.9
Pillow==10.2.0